Fix signed/unsigned mismatch for linux
BUG=none
TEST=none

Review URL: http://codereview.chromium.org/261004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@28319 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/flip/flip_session_pool.cc
@@ -11,7 +11,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace net {
 
 // The maximum number of sessions to open to a single domain.
-const int kMaxSessionsPerDomain = 1;
+const size_t kMaxSessionsPerDomain = 1;
 
 scoped_ptr<FlipSessionPool::FlipSessionsMap> FlipSessionPool::sessions_;
